+++ include/core/doublebufferedallocator.h
@@ -12,10 +12,10 @@ namespace fissura
 		DoubleBufferedAllocator(StackAllocator& stackA, StackAllocator& stackB);
 		~DoubleBufferedAllocator();
 
-		void* allocate(u32 size, u8 alignment);
+		void* allocate(size_t size, u8 alignment);
 		void deallocate(void* p);
 
-		u32 getTotalUsedMemory() const;
+		size_t getTotalUsedMemory() const;
 		u32 getTotalNumAllocations() const;
 		bool canDeallocate() const { return false; }
